web_service/requirements.txt
Flask==2.1.1
pandas==1.3.3
scikit-learn==0.24.2
mysql-connector-python==8.0.28
